[TTS][refactor] Part 2 - nemo.colletions.tts.parts
