+++ nemo/collections/tts/losses/spectrogram_enhancer_losses.py
@@ -41,7 +41,7 @@ import torch.nn.functional as F
 from einops import rearrange
 from torch.autograd import grad as torch_grad
 
-from nemo.collections.tts.helpers.helpers import mask_sequence_tensor
+from nemo.collections.tts.parts.utils.helpers import mask_sequence_tensor
 
 
 class GradientPenaltyLoss(torch.nn.Module):
