+++ nemo/collections/tts/losses/tacotron2loss.py
@@ -14,7 +14,7 @@
 
 import torch
 
-from nemo.collections.tts.helpers.helpers import get_mask_from_lengths
+from nemo.collections.tts.parts.utils.helpers import get_mask_from_lengths
 from nemo.core.classes import Loss, typecheck
 from nemo.core.neural_types.elements import LengthsType, LogitsType, LossType, MelSpectrogramType
 from nemo.core.neural_types.neural_type import NeuralType
