+++ nemo/collections/tts/modules/tacotron2.py
@@ -16,8 +16,8 @@ import torch
 from torch.autograd import Variable
 from torch.nn import functional as F
 
-from nemo.collections.tts.helpers.helpers import get_mask_from_lengths
 from nemo.collections.tts.modules.submodules import Attention, ConvNorm, LinearNorm, Prenet
+from nemo.collections.tts.parts.utils.helpers import get_mask_from_lengths
 from nemo.core.classes import NeuralModule, typecheck
 from nemo.core.neural_types.elements import (
     EmbeddedTextType,
