+++ nemo/collections/tts/modules/waveglow.py
@@ -15,8 +15,8 @@ from typing import Tuple
 
 import torch
 
-from nemo.collections.tts.helpers.helpers import OperationMode, remove, split_view
 from nemo.collections.tts.modules.submodules import Invertible1x1Conv, WaveNet
+from nemo.collections.tts.parts.utils.helpers import OperationMode, remove, split_view
 from nemo.core.classes import Exportable, NeuralModule, typecheck
 from nemo.core.neural_types.elements import (
     AudioSignal,
