+++ scripts/dataset_processing/tts/generate_mels.py
@@ -33,7 +33,10 @@ from joblib import Parallel, delayed
 from tqdm import tqdm
 
 from nemo.collections.tts.models import FastPitchModel
-from nemo.collections.tts.torch.helpers import BetaBinomialInterpolator, beta_binomial_prior_distribution
+from nemo.collections.tts.parts.utils.tts_dataset_utils import (
+    BetaBinomialInterpolator,
+    beta_binomial_prior_distribution,
+)
 from nemo.utils import logging
