+++ scripts/dataset_processing/tts/resynthesize_dataset.py
@@ -63,10 +63,10 @@ import torch
 from omegaconf import DictConfig, OmegaConf
 from tqdm import tqdm
 
-from nemo.collections.tts.data.data_utils import read_manifest, write_manifest
-from nemo.collections.tts.helpers.helpers import process_batch, to_device_recursive
 from nemo.collections.tts.models import FastPitchModel
 from nemo.collections.tts.models.base import SpectrogramGenerator
+from nemo.collections.tts.parts.utils.helpers import process_batch, to_device_recursive
+from nemo.collections.tts.parts.utils.tts_dataset_utils import read_manifest, write_manifest
 
 
 def chunks(iterable: Iterable, size: int) -> Iterator[List]:
