+++ scripts/ssl_tts/make_supdata.py
@@ -29,7 +29,7 @@ from tqdm import tqdm
 
 from nemo.collections.asr.parts.preprocessing.segment import AudioSegment
 from nemo.collections.tts.models import ssl_tts
-from nemo.collections.tts.torch.helpers import get_base_dir
+from nemo.collections.tts.parts.utils.tts_dataset_utils import get_base_dir
 from nemo.core.classes import Dataset
 from nemo.utils import logging
