+++ tests/collections/tts/data/test_audio_trimming.py
@@ -15,7 +15,10 @@
 import numpy as np
 import pytest
 
-from nemo.collections.tts.data.audio_trimming import get_start_and_end_of_speech_frames, pad_sample_indices
+from nemo.collections.tts.parts.preprocessing.audio_trimming import (
+    get_start_and_end_of_speech_frames,
+    pad_sample_indices,
+)
 
 
 class TestAudioTrimming:
