+++ tests/collections/tts/test_spectrogram_enhancer.py
@@ -17,8 +17,8 @@ import torch
 from einops import rearrange
 from omegaconf import DictConfig
 
-from nemo.collections.tts.helpers.helpers import mask_sequence_tensor
 from nemo.collections.tts.models import SpectrogramEnhancerModel
+from nemo.collections.tts.parts.utils.helpers import mask_sequence_tensor
 
 
 @pytest.fixture
